active card style added

+++ src/components/taskBoard/taskboard.js
@@ -3,13 +3,14 @@ import { connect } from 'react-redux';
 import { Button, Row, Col, ButtonGroup, ListGroup, ListGroupItem } from 'reactstrap';
 // import {Link} from 'react-router-dom'
 import SunEditor from 'suneditor-react';
-import { RiAddCircleFill, RiRefreshFill, RiEditCircleFill, RiSaveFill, RiDeleteBin2Line} from 'react-icons/ri';
+import { RiAddCircleFill, RiRefreshFill, RiSaveFill, RiDeleteBin2Line} from 'react-icons/ri';
 
 import editorConfig from './editor.config'
 import { fetchTasksboardAsync } from '../../store/reducer/asyncReducer';
 import Board from './board/board'
 import axios from '../../http/axios';
 import WithModal from '../hoc/withModal';
+import WithTooltip from '../hoc/withTooltip';
 import BoardForm from './form/boardForm';
 
 const Notebook = (props) => {
@@ -20,11 +21,13 @@ const Notebook = (props) => {
   useEffect(() => {
     if(!props.tasksboard.length){
       props.fetchTaskboard()
+    } else {
+      clickTaskHandler(props.tasksboard[0])
     }
     return () => {
-      console.log('Notebook removed from view!')
+      console.log('Notebook js cleanup')
     }
-  }, [])
+  }, [props.tasksboard.length])
   
   const fetchTaskboardHandler = () => {
     props.fetchTaskboard()
@@ -50,7 +53,8 @@ const Notebook = (props) => {
   }
 
   const deleteTaskboardHandler = () => {
-    if(currentTask){
+    const result = window.confirm('Are you sure you want to delete?');
+    if(currentTask && result){
       axios.delete(`taskboard?taskboardId=${currentTask._id}`).then(
         res=>console.log(res.data)
       ).catch(
@@ -64,32 +68,33 @@ const Notebook = (props) => {
       <Col sm="6" md="4" lg="3" color="secondary" className="p-0 taskListContainer bg-light">
         <div className="text-end w-100">
           <ButtonGroup className="w-auto mr-auto">
-            <Button color="ternary" className="border-radius-0" onClick={props.toggle}>
-              <RiAddCircleFill ></RiAddCircleFill>
-            </Button>
-            <Button color="ternary" className="border-radius-0" onClick={fetchTaskboardHandler}>
-              <RiRefreshFill></RiRefreshFill>
-            </Button>
+            <WithTooltip>
+              <Button color="ternary" className="border-radius-0" data-tip="Add" onClick={props.toggle}>
+                <RiAddCircleFill ></RiAddCircleFill>
+              </Button>
+              <Button color="ternary" className="border-radius-0" data-tip="Refresh" onClick={fetchTaskboardHandler}>
+                <RiRefreshFill></RiRefreshFill>
+              </Button>
+            </WithTooltip>
           </ButtonGroup>
         </div>
         <div className="heightFixTaskList">
           <ListGroup>
-            {props.tasksboard.map((item) => <ListGroupItem key={item._id}><Board clicked={clickTaskHandler} data={item}></Board></ListGroupItem>)}
+            {props.tasksboard.map((item) => <ListGroupItem key={item._id}><Board clicked={clickTaskHandler} selected={currentTask} data={item}></Board></ListGroupItem>)}
           </ListGroup>
         </div>
       </Col>
       <Col sm="6" md="8" lg="9" color="secondary" className="p-0 editorContainer">
         <div className="text-end w-100 bg-light ">
           <ButtonGroup className="w-auto mr-auto">
-            <Button color="ternary" className="border-radius-0" onClick={saveTaskBoardHandler}>
-              <RiSaveFill></RiSaveFill>
-            </Button>
-            <Button color="ternary" className="border-radius-0">
-              <RiEditCircleFill></RiEditCircleFill>
-            </Button>
-            <Button color="ternary" className="border-radius-0" onClick={deleteTaskboardHandler}>
-              <RiDeleteBin2Line></RiDeleteBin2Line>
-            </Button>
+            <WithTooltip>
+              <Button color="ternary" className="border-radius-0" data-tip="Save" onClick={saveTaskBoardHandler}>
+                <RiSaveFill></RiSaveFill>
+              </Button>
+              <Button color="ternary" className="border-radius-0" data-tip="Delete" onClick={deleteTaskboardHandler}>
+                <RiDeleteBin2Line></RiDeleteBin2Line>
+              </Button>
+            </WithTooltip>
           </ButtonGroup>
         </div>
         <SunEditor enable={true} disable={false} show={true} showToolbar={true} setOptions={editorConfig} ref={editorRef}/>
